arrumando a modelagem do dominio

+++ src/main/java/br/com/caelum/comidinhas/restaurante/Avaliacao.java
@@ -1,6 +1,7 @@
 package br.com.caelum.comidinhas.restaurante;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -8,7 +9,7 @@ public class Avaliacao {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long nota;
+    private BigDecimal nota;
     @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
     @JoinColumn(name = "avaliacao_id")
     private List<AvaliacaoComentarios> avaliacaoComentarios;
